backend/requirements.txt
fastapi
uvicorn
paypal-checkout-serversdk
python-dotenv